Extract twitter delegate into a separate file since this is a shared functionality now
