+++ trySwift/UIViewControllerExtension.swift
@@ -23,27 +23,6 @@ extension UIViewController: SFSafariViewControllerDelegate {
     }
 }
 
-extension UIViewController: TwitterFollowDelegate {
-    
-    func followUser(_ username: String) {
-        var applicationOpened = false
-        let application = UIApplication.shared
-        for twitterURL in Twitter.urls(forUsername: username) {
-            if let url = URL(string: twitterURL) , application.canOpenURL(url) && !applicationOpened {
-                application.open(url, options: [String:Any](), completionHandler: nil)
-                applicationOpened = true
-                break
-            }
-        }
-        
-        if !applicationOpened {
-            if let twitterURL = URL(string: "https://twitter.com/\(username)") {
-                openSafariViewController(withURL: twitterURL)
-            }
-        }
-    }
-}
-
 extension UIViewController: MFMailComposeViewControllerDelegate {
     
     func sendMail(withConfiguration configuration: MailConfiguration) {
